[bugfix] change save unsigned int to long long
1. cosvis = 8 for superpoint and 5 for orb
2. increase parallax
3. remove origin model of superpoint
4. use brute force instead of dbow for sfm match
5. draw boundingbox plane color

+++ include/ORBSLAM3/Atlas.h
@@ -141,8 +141,8 @@ public:
     bool MappointInBoundingbox(const cv::Mat &pos_mat);
     void GetBoundingBoxCoordsRange();
     bool WriteToMemoryFor3DObject(
-        const unsigned int &mem_size, char *mem, const bool is_superpoint);
-    unsigned int
+        const long long &mem_size, char *mem, const bool is_superpoint);
+    long long
     GetMemSizeFor3DObject(const std::string &version, const bool is_superpoint);
     void SetScanBoundingbox_W(const std::vector<Eigen::Vector3d> &boundingbox);
     std::set<Map *> mspMaps;
